vehicle-signal-server-spec/vsss-get-simple-test.html

<!DOCTYPE HTML>
<meta charset="utf-8">
<title>Vehicle Signal Server Spec Test: get method tests</title>
<link rel="help" href="https://www.w3.org/TR/vehicle-information-service/#idl-def-getrequest">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./server-setting.js"></script>

<p> Run VSSS server before starting test </p>

<div id="log"></div>

<script>
var t = async_test("Get method test");

var vehicle  = new WebSocket("ws://"+VSSS_HOST+":"+VSSS_PORT);

vehicle.onopen = function() {
  // request a signal
  var reqId = "reqid-12345";
  var path = "Signal.Drivetrain.Transmission.Speed";

  // process messages from the server
  vehicle.onmessage = function(event) {

    t.step(function(){
      var msg = JSON.parse(event.data);
      console.log("event.data: " + event.data);
      if(msg.requestId == reqId && msg.value){
        assert_true(true, "OK: "+path + ": value = " + msg.value);
      }
      vehicle.close();
      t.done();
    });
  }
  vehicle.send('{"action":"get","path":"'+ path +'" ,"requestId":"'+reqId+'"}');
}
</script>
